pyproject.toml
[project]
name = "ct-agent"
version = "0.1.0"
description = "Copy trading agent for telegram"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]

[tool.uv.workspace]
members = [
     "modules/ct_agent_broker_ex_module",
     "modules/ct_agent_nlp_parser_module",
     "modules/ct_agent_tg_bot_module",
     "shared",
]
